Display request failures as errors in the status page
When a request fails before the `kubectl apply` phase we should still chnage the
Waybill.Status.LastRun.Success to mark it as a failure and display it as such
in the status page.
This will produce a failed run for such cases, with empty vaules on fiels
populated after `kubectl apply`. The status page should omit commit row and
outputs if not specified.
Also, we should always default to a "green" panel for successfully applied
namespaces.