Set gitconfig for strongbox under temp home dir
Problem:
Repository clone happens using a global gitconfig for the current user (root)
that is already configured for strongbox in the Dockrfike we use to build
images. `kustomize` runs as part of Apply function when the controller has
created a temporary home dir to use in order to set ssh configs. If the
kustomize remote base contains strongbox encrypted files, then the temporary
home gitconfig is not configured to cover us so that we can decrypt successfully
on pull and remote base encrypted files remain encrypted.

This is trying to address the above by creating a temporary gitconfig under the
temp home per apply and sets STRONGBOX_HOME environment variable to point to
the temp home as well. More in particular:
- Creates an interface for strongbox and adds it to the Runner.
- Mocks the function that tries to exec strongbox in order to configure git for
  testing.
- Explicitly sets STRONGBOX_HOME before running apply command